Still debugging merged a2a

+++ cps_pp/include/alg/enum_int.h
@@ -19,7 +19,7 @@ typedef double Float;
 
 typedef double IFloat;
 
-typedef uint32_t Pointer;
+typedef uint64_t Pointer;
 
 enum DirType {
 	DIR_X = 0,
